+++ CHANGELOG.md
@@ -11,6 +11,8 @@ next release / 2014-02-??
 * removed `sign()` and `verify()`, methods can be accessed from [ecdsa](https://github.com/cryptocoinjs/ecdsa)
 * added `privateExportKey` 
 * removed `getExportedPrivateKey`, note that `getExportedPrivateKey` was essentially just a way to get WIF
+* removed `decodeString()`, use package [coinstring][coinstring] in its place
+* removed `getBitcoinAddress()`, use package [coinstring][coinstring] in its place
 
 0.3.0 / 2014-02-03
 ------------------
@@ -40,3 +42,5 @@ next release / 2014-02-??
 0.0.1 / 2013-11-12
 ------------------
 * initial release
+
+[coinstring]: https://github.com/cryptocoinjs/coinstring